+++ class/GaiaAlpha/Controller/CmsController.php
@@ -156,12 +156,12 @@ class CmsController extends BaseController
         $this->jsonResponse(['url' => $mediaUrl]);
     }
 
-    public function registerRoutes(\GaiaAlpha\Router $router)
+    public function registerRoutes()
     {
-        $router->add('GET', '/api/cms/pages', [$this, 'index']);
-        $router->add('POST', '/api/cms/pages', [$this, 'create']);
-        $router->add('PATCH', '/api/cms/pages/(\d+)', [$this, 'update']);
-        $router->add('DELETE', '/api/cms/pages/(\d+)', [$this, 'delete']);
-        $router->add('POST', '/api/cms/upload', [$this, 'upload']);
+        \GaiaAlpha\Router::add('GET', '/api/cms/pages', [$this, 'index']);
+        \GaiaAlpha\Router::add('POST', '/api/cms/pages', [$this, 'create']);
+        \GaiaAlpha\Router::add('PATCH', '/api/cms/pages/(\d+)', [$this, 'update']);
+        \GaiaAlpha\Router::add('DELETE', '/api/cms/pages/(\d+)', [$this, 'delete']);
+        \GaiaAlpha\Router::add('POST', '/api/cms/upload', [$this, 'upload']);
     }
 }