+++ class/GaiaAlpha/Router.php
@@ -17,36 +17,7 @@ class Router
         ];
     }
 
-    public static function loadControllers()
-    {
-        $rootDir = Env::get('root_dir');
-        // Dynamically Init Controllers
-        $controllers = [];
-        foreach (glob($rootDir . '/class/GaiaAlpha/Controller/*Controller.php') as $file) {
-            $filename = basename($file, '.php');
-            if ($filename === 'BaseController')
-                continue;
 
-            $key = strtolower(str_replace('Controller', '', $filename));
-            $className = "GaiaAlpha\\Controller\\$filename";
-
-            if (class_exists($className)) {
-                $controller = new $className();
-                if (method_exists($controller, 'init')) {
-                    $controller->init();
-                }
-                $controllers[$key] = $controller;
-            }
-        }
-
-        Env::set('controllers', $controllers);
-
-        // Pass an instance to satisfy type hints, even though methods are static
-        $routerInstance = new self();
-        foreach ($controllers as $controller) {
-            $controller->registerRoutes($routerInstance);
-        }
-    }
 
     public static function dispatch(string $method, string $uri)
     {